requirements.txt
"pathway[xpack-llm]"
requests
streamlit
openai
python-dotenv